Refactor rendering to seperate concerns
