+++ src/app/components/Headline/index.jsx
@@ -1,0 +1,9 @@
+import styled from 'styled-components';
+
+const Headline = styled.h1`
+  color: #222;
+  font-family: ReithSans, Arial, Helvetica, freesans, sans-serif;
+  font-size: 2em;
+`;
+
+export default Headline;
